@@ -8,8 +8,8 @@
   "types": "index.d.ts",
   "sideEffects": false,
   "scripts": {
-    "dev": "parcel demo/index.html -d .devserver --no-cache",
-    "build": "rollup --config",
+    "dev": "parcel serve demo/index.html",
+    "build": "rollup --config --bundleConfigAsCjs",
     "test": "jest --coverage",
     "lint": "eslint \"{src,test}/**/*.js\"",
     "format": "prettier --write \"{src,test,demo}/**/*.{js,ts,json,css,md}\""
@@ -28,6 +28,7 @@
     "index.d.ts"
   ],
   "jest": {
+    "testEnvironment": "jsdom",
     "setupFilesAfterEnv": [
       "<rootDir>test/setup.js",
       "@testing-library/jest-dom/extend-expect"
@@ -52,53 +53,50 @@
     ]
   },
   "dependencies": {
+    "@babel/core": "^7.22.0",
+    "react-is": "^18.2.0",
     "tippy.js": "^6.3.1"
   },
+  "devDependencies": {
+    "@babel/core": "^7.22.0",
+    "@babel/eslint-parser": "^7.22.0",
+    "@babel/plugin-syntax-jsx": "^7.22.0",
+    "@babel/preset-env": "^7.22.0",
+    "@babel/preset-react": "^7.22.0",
+    "@rollup/plugin-babel": "^6.0.0",
+    "@rollup/plugin-node-resolve": "^15.0.0",
+    "@rollup/plugin-replace": "^5.0.0",
+    "@rollup/plugin-terser": "^0.4.3",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@types/react": "^19.0.0",
+    "@types/react-dom": "^19.0.0",
+    "babel-plugin-annotate-pure-calls": "^0.4.0",
+    "babel-plugin-styled-components": "^2.1.4",
+    "core-js": "^3.30.0",
+    "eslint": "^8.40.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "framer-motion": "^10.12.0",
+    "husky": "^8.0.0",
+    "jest": "^29.5.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "lint-staged": "^13.2.2",
+    "parcel": "^2.8.3",
+    "prettier": "^2.8.8",
+    "process": "^0.11.10",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-spring": "^9.7.1",
+    "rollup": "^3.21.0",
+    "styled-components": "^5.3.10",
+    "typescript": "^5.0.4"
+  },
   "peerDependencies": {
+    "@babel/core": "^7.0.0",
     "react": ">=16.8",
     "react-dom": ">=16.8"
   },
-  "devDependencies": {
-    "@babel/core": "^7.8.7",
-    "@babel/preset-env": "^7.8.7",
-    "@babel/preset-react": "^7.0.0",
-    "@testing-library/jest-dom": "^5.12.0",
-    "@testing-library/react": "^11.2.7",
-    "@types/react": "^16.8.2",
-    "babel-eslint": "^10.0.1",
-    "babel-jest": "^25.1.0",
-    "babel-plugin-annotate-pure-calls": "^0.4.0",
-    "core-js": "^3.6.4",
-    "eslint": "^5.14.1",
-    "eslint-config-prettier": "^3.6.0",
-    "eslint-plugin-react": "^7.12.4",
-    "eslint-plugin-react-hooks": "^1.7.0",
-    "framer-motion": "^1.10.3",
-    "husky": "^1.3.1",
-    "jest": "^24.1.0",
-    "lint-staged": "^8.1.0",
-    "parcel": "^1.12.3",
-    "prettier": "^1.16.1",
-    "react": "^16.8.1",
-    "react-dom": "^16.8.1",
-    "react-spring": "^8.0.27",
-    "rollup": "^1.14.3",
-    "rollup-plugin-babel": "^4.3.2",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-replace": "^2.2.0",
-    "rollup-plugin-terser": "^5.2.0",
-    "styled-components": "^5.0.1",
-    "typescript": "^3.6.3"
-  },
-  "directories": {
-    "test": "test"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/atomiks/tippyjs-react.git"
-  },
-  "bugs": {
-    "url": "https://github.com/atomiks/tippyjs-react/issues"
-  },
-  "homepage": "https://github.com/atomiks/tippyjs-react#readme"
+  "packageManager": "yarn@3.5.1+sha256.64c0b63e5fa0e21f354b5eff7c9a1b486f32046bfc3283670569e3c6a9cad102"
 }
